client/requirements.txt
pyzmq
python-dotenv
pygame